+++ security/src/main/scala/pl/touk/nussknacker/ui/security/oauth2/OAuth2AuthenticationResources.scala
@@ -24,7 +24,8 @@ class OAuth2AuthenticationResources(override val name: String, realm: String, se
     configuration.authSeverPublicKey.map(CertificatesAndKeys.textualRepresentationOfPublicKey),
     configuration.idTokenNonceVerificationRequired,
     configuration.implicitGrantEnabled,
-    configuration.anonymousUserRole.isDefined
+    configuration.anonymousUserRole.isDefined,
+    configuration.jwt.flatMap(_.audience)
   )
 
   val anonymousUserRole: Option[String] = configuration.anonymousUserRole